requirements.txt
dspy>=3.1.2
rich
python-dotenv
